+++ src/features/laundry-item/laundry-item.validation.ts
@@ -3,8 +3,8 @@ import z from "zod";
 export class LaundryItemValidation {
   static readonly QUERY = {
     getLaundryItems: z.object({
-      page: z.coerce.number().positive().optional(),
-      pageSize: z.coerce.number().positive().optional(),
+      page: z.coerce.number().int().positive().optional(),
+      pageSize: z.coerce.number().int().positive().optional(),
       search: z.string().trim().optional(),
       sortBy: z.enum(["name", "createdAt"]).optional(),
       sortOrder: z.enum(["asc", "desc"]).optional(),
